openapis/requirements.txt
# dvc==0.22.0
pytest==4.0.1
PyYAML
